implement design feedback and refactor files
